KS-16-4 deleted class-files
